+++ src/api/FleetManager/FleetManager.Domain/VehicleUsages/FuelExpense.cs
@@ -5,12 +5,12 @@
         public Guid Id { get; private set; }
         public Guid VehicleUsageId { get; private set; }
         public Guid RouteId { get; private set; }
-        public decimal Amount { get; private set; }
-        public decimal Liters { get; private set; }
+        public double Amount { get; private set; }
+        public double Liters { get; private set; }
         public FuelType FuelType { get; private set; }
         public DateTime Date { get; private set; }
 
-        public FuelExpense(Guid vehicleUsageId, Guid routeId, decimal amount, decimal liters, FuelType fuelType)
+        public FuelExpense(Guid vehicleUsageId, Guid routeId, double amount, double liters, FuelType fuelType)
         {
             Id = Guid.NewGuid();
             VehicleUsageId = vehicleUsageId;
