+++ src/api/FleetManager/FleetManager.Domain/Vehicles/Repair.cs
@@ -6,9 +6,9 @@
         public Guid VehicleId { get; private set; }
         public DateTime Date { get; private set; }
         public string Description { get; private set; }
-        public decimal Cost { get; private set; }
+        public double Cost { get; private set; }
 
-        public Repair(Guid vehicleId, DateTime date, string description, decimal cost)
+        public Repair(Guid vehicleId, DateTime date, string description, double cost)
         {
             Id = Guid.NewGuid();
             VehicleId = vehicleId;
